comment 일부수정

+++ controllers/diaries.controllers.js
@@ -13,35 +13,41 @@ class DiaryController {
   createDiary = async (req, res) => {
     const { userId } = req.params;
     const { content, diaryNo } = req.body;
-    // const { name } = res.locals.user;
-    // const userInfo = res.locals.user;
+    const { name } = res.locals.user;
+    const userInfo = res.locals.user;
 
-    if (!content) {
-      res.status(400).json({ message: '내용을 입력해주세요!' });
+    try {
+      if (!content) {
+        res.status(400).json({ message: '내용을 입력해주세요!' });
+      }
+      // 본인 이외의 사람이 다이어리 작성시 예외처리
+      if (userInfo.userId !== Number(userId)) {
+        throw new Error('작성 권한이 없습니다.');
+      }
+      // 파일이 있으면 key값으로 이름을 정하고 없으면 null
+      const imageFileName = req.file ? req.file.key : null;
+      // imageFileName에 파일명이 들어 갔으면 s3 url주소 추가
+      const dirImg = imageFileName
+        ? process.env.S3_STORAGE_URL + imageFileName
+        : null;
+
+      const createDiaryData = await this.diaryController.createDiary(
+        userId,
+        // name,
+        dirImg,
+        content,
+        diaryNo
+      );
+      res.status(200).json({ msg: '다이어리 작성완료!' });
+    } catch (err) {
+      res.status(400).json({ err: err.message });
     }
-    // if (userInfo.userId !== Number(userId)) {
-    //   throw new Error('작성 권한이 없습니다..');
-    // }
-    // 파일이 있으면 key값으로 이름을 정하고 없으면 null
-    const imageFileName = req.file ? req.file.key : null;
-    // imageFileName에 파일명이 들어 갔으면 s3 url주소 추가
-    const dirImg = imageFileName
-      ? process.env.S3_STORAGE_URL + imageFileName
-      : null;
-
-    const createDiaryData = await this.diaryController.createDiary(
-      userId,
-      // name,
-      dirImg,
-      content,
-      diaryNo
-    );
-    res.status(200).json({ data: createDiaryData });
   };
 
   updateDiary = async (req, res) => {
     const { diaryId, userId } = req.params;
     const { content } = req.body;
+    const userInfo = res.locals.user;
 
     // 수정사항에 이미지 파일이 있으면 key값으로 이름 정해주고 없으면 Null
     const imageFileName = req.file ? req.file.key : null;
@@ -49,23 +55,39 @@ class DiaryController {
     const dirImg = imageFileName
       ? process.env.S3_STORAGE_URL + imageFileName
       : undefined;
+    try {
+      // 본인 이외의 사람이 다이어리 수정시 예외처리
+      if (userInfo.userId !== Number(userId)) {
+        throw new Error('수정 권한이 없습니다.');
+      }
 
-    const updateDiaryData = await this.diaryController.updateDiary(
-      diaryId,
-      // userId,
-      dirImg,
-      content
-    );
-    res.status(200).json({ data: updateDiaryData });
+      const updateDiaryData = await this.diaryController.updateDiary(
+        diaryId,
+        // userId,
+        dirImg,
+        content
+      );
+      res.status(200).json({ data: updateDiaryData });
+    } catch (err) {
+      res.status(400).json({ err: err.message });
+    }
   };
 
   deleteDiary = async (req, res) => {
     const { diaryId } = req.params;
-    // const { userId } = req.params;
+    const { userId } = req.params;
+    const userInfo = res.locals.user;
 
-    const deleteDiaryData = await this.diaryController.deleteDiary(diaryId);
-    // const diaries = await this.diaryController.findAllDiary(userId);
-    res.status(200).json({ data: deleteDiaryData });
+    try {
+      // 본인 이외의 사람이 다이어리 삭제시 예외처리
+      if (userInfo.userId !== Number(userId)) {
+        throw new Error('삭제 권한이 없습니다.');
+      }
+      const deleteDiaryData = await this.diaryController.deleteDiary(diaryId);
+      res.status(200).json({ msg: '삭제되었습니다!' });
+    } catch (err) {
+      res.status(400).json({ err: err.message });
+    }
   };
 }
